% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation.R
\name{cash}
\alias{cash}
\title{CASH clustering algorithm}
\usage{
cash(data_matrix, minpts, maxlevel, jitter, adjust = NA, mindim = NA)
}
\arguments{
\item{data_matrix}{numeric Matrix containing data the clusters are
calculated for. Rows are treated as observations, columns as features.}

\item{minpts}{Number. Threshold for minimum number of points in a cluster.
Has to be >= 1.}

\item{maxlevel}{Number. Maximum level for splitting the hypercube.
Has to be >= 1.}

\item{jitter}{Number. Maximum jitter for distance values.
Has to be > 0.}

\item{adjust}{Boolean. Apply adjustment heuristic for interval choosing.}

\item{mindim}{Number. Minimum dimensionality of the subspaces to be found.}
}
\value{
List of clusters. The indices of each cluster belong to the indices
           in the given original data_matrix
}
\description{
\code{cash} returns a list of clusters for the given data_matrix, formed according to CASH.
}
\examples{
data_matrix <- matrix(c(1:30), nrow=10, ncol=3)
result      <- cash(data_matrix, 5, 10, 1, adjust = TRUE)
for(cluster_index in c(1:length(result))) {
    print(paste('Cluster: ', cluster_index))
    print(paste('Indices: ', paste(result[[cluster_index]], collapse=',')))
}


}
\seealso{
\url{https://elki-project.github.io/releases/release0.7.5/javadoc/de/lmu/ifi/dbs/elki/algorithm/clustering/correlation/CASH.html}
  for ELKI documentation.
}
