% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_lof.R
\name{inflo}
\alias{inflo}
\title{INFLO Outlier score calculation.}
\usage{
inflo(data_matrix, k, m = NA)
}
\arguments{
\item{data_matrix}{numeric Matrix containing data the outlier score is
calculated for. Rows are treated as observations, columns as features.}

\item{k}{Number. Neighbourhood-size used to calculate outlier scores.}

\item{m}{Floating point number. Is used as pruning threshold and has to be
greater than 0.0. If NA, ELKI's default is used (1.0).}
}
\value{
List of outlier scores. The score at position x belongs to the
  observation given in row x of the original data_matrix.
}
\description{
\code{inflo} returns the INFLO Outlier score for every observation in the
given data_matrix. INFLO is short for Influence Outlierness using Symetric
Relationship. LOF in combination with a reverse kNN is used to calculate
outlier score.
}
\examples{
data_matrix <- matrix(c(1:30), nrow=10, ncol=3)
result      <- inflo(data_matrix, 3)
for(index in c(1:10)) {
    print(paste('Observation:', paste(data_matrix[index,], collapse=',')))
    print(paste('Score:',       result[index]))
}


}
\seealso{
\url{https://elki-project.github.io/releases/release0.7.5/javadoc/de/lmu/ifi/dbs/elki/algorithm/outlier/lof/INFLO.html}
for ELKI documentation.
}
