% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_lof.R
\name{ldf}
\alias{ldf}
\title{LDF Outlier score calculation.}
\usage{
ldf(data_matrix, k, h = 1, c = 0.1)
}
\arguments{
\item{data_matrix}{numeric Matrix containing data the outlier score is
calculated for. Rows are treated as observations, columns as features.}

\item{k}{Number. Neighbourhood-size used to calculate outlier scores.}

\item{h}{Number. Bandwidth scaling factor. Defaults to ELKI's default (1).}

\item{c}{Number. Scaling constant to limit value range to 1/c. Defaults to ELKI's default (0.1).}
}
\value{
List of outlier scores. The score at position x belongs to the
  observation given in row x of the original data_matrix.
}
\description{
\code{ldf} returns the LDF Outlier score for every observation in the given
data_matrix. Kernel density estimation in
combination with the reachability concept of LOF is used to calculate outlier score.
}
\examples{
data_matrix <- matrix(c(1:30), nrow=10, ncol=3)
result      <- ldf(data_matrix, 3)
for(index in c(1:10)) {
    print(paste('Observation:', paste(data_matrix[index,], collapse=',')))
    print(paste('Score:',       result[index]))
}


}
\seealso{
\url{https://elki-project.github.io/releases/release0.7.5/javadoc/de/lmu/ifi/dbs/elki/algorithm/outlier/lof/LDF.html}
for ELKI documentation.
}
