% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_lof.R
\name{loop}
\alias{loop}
\title{LoOP Outlier score calculation}
\usage{
loop(data_matrix, k_reach, k_comp = k_reach, lambda = NA)
}
\arguments{
\item{data_matrix}{numeric Matrix containing data the outlier score is
calculated for. Rows are treated as observations, columns as features.}

\item{k_reach}{Number. Neighbourhood-size used to calculate outlier scores.}

\item{k_comp}{Number. Neighbourhood-size used to calculate outlier scores.
Defaults to k_reach}

\item{lambda}{Number. If NA, ELKI's default is used (2.0).}
}
\value{
List of outlier scores. The score at position x belongs to the
  observation given in row x of the original data_matrix.
}
\description{
\code{loop} returns the LoOP Outlier score for every observation in the given
data_matrix. LoOP is short for Local Outlier Probabilities. LOF in
combination with statistical methods is used to calculate outlier score.
}
\examples{
data_matrix <- matrix(c(1:30), nrow=10, ncol=3)
result      <- loop(data_matrix, 3)
for(index in c(1:10)) {
    print(paste('Observation:', paste(data_matrix[index,], collapse=',')))
    print(paste('Score:',       result[index]))
}


}
\seealso{
\url{https://elki-project.github.io/releases/release0.7.5/javadoc/de/lmu/ifi/dbs/elki/algorithm/outlier/lof/LoOP.html}
for ELKI documentation.
}
